Enhance Makefiles and fix integer type usage in code.

+++ current/dox/connect_ethernet.c
@@ -1,19 +1,18 @@
 #include "urg_sensor.h"
 
-
-int main(void)
+int32_t main(void)
 {
     urg_t urg;
-    int ret;
-// \~japanese ƒC[ƒT[ƒlƒbƒgÚ‘±‚Å‚ÌƒZƒ“ƒT‚Æ‚ÌÚ‘±‚Æ‹——£ƒf[ƒ^‚Ìæ“¾
-// \~english Connects to the sensor via Ethernet and receives range data
+    int32_t ret;
+    // \~japanese ã‚¤ãƒ¼ã‚µãƒ¼ãƒãƒƒãƒˆæ¥ç¶šã§ã®ã‚»ãƒ³ã‚µã¨ã®æ¥ç¶šã¨è·é›¢ãƒ‡ãƒ¼ã‚¿ã®å–å¾—
+    // \~english Connects to the sensor via Ethernet and receives range data
 
-const char connect_address[] = "192.168.0.10";
-const long connect_port = 10940;
+    const char connect_address[] = "192.168.0.10";
+    const int32_t connect_port = 10940;
 
-// \~japanese ƒZƒ“ƒT‚É‘Î‚µ‚ÄÚ‘±‚ğs‚¤B
-// \~english Connects to the sensor
-ret = urg_open(&urg, URG_ETHERNET, connect_address, connect_port);
-// \todo check error code
-return 0;
+    // \~japanese ã‚»ãƒ³ã‚µã«å¯¾ã—ã¦æ¥ç¶šã‚’è¡Œã†ã€‚
+    // \~english Connects to the sensor
+    ret = urg_open(&urg, URG_ETHERNET, connect_address, connect_port);
+    // \todo check error code
+    return 0;
 }
