+++ current/include/cpp/ticks.h
@@ -1,6 +1,8 @@
 #ifndef QRK_TICKS_H
 #define QRK_TICKS_H
 
+#include <stdint.h>
+
 /*!
   \file
   \~japanese
@@ -13,10 +15,10 @@
   $Id$
 */
 
-
 namespace qrk
 {
-    long ticks(void);
+  //! \~japanese 現在のタイムスタンプを取得する  \~english Gets the current timestamp
+  int32_t ticks(void);
 }
 
 #endif /* !QRK_TICKS_H */
