+++ current/samples/cpp/get_distance_intensity.cpp
@@ -16,18 +16,17 @@
 using namespace qrk;
 using namespace std;
 
-
 namespace
 {
-    void print_data(const Urg_driver& urg,
-                    const vector<long>& data,
-                    const vector<unsigned short>& intensity,
-                    long time_stamp)
+    void print_data(const Urg_driver &urg,
+                    const vector<int32_t> &data,
+                    const vector<unsigned short> &intensity,
+                    int32_t time_stamp)
     {
 #if 1
         // \~japanese 前方のデータのみを表示
         // \~english Shows only the front step
-        int front_index = urg.step2index(0);
+        int32_t front_index = urg.step2index(0);
         cout << data[front_index] << " [mm], "
              << intensity[front_index] << " [1], ("
              << time_stamp << " [msec])" << endl;
@@ -39,15 +38,15 @@ namespace
         // \~english Prints the range/intensity values for all the measurement points
         size_t data_n = data.size();
         cout << "# n = " << data_n << ", timestamp = " << time_stamp << endl;
-        for (size_t i = 0; i < data_n; ++i) {
+        for (size_t i = 0; i < data_n; ++i)
+        {
             cout << i << ", " << data[i] << ", " << intensity[i] << endl;
         }
 #endif
     }
 }
 
-
-int main(int argc, char *argv[])
+int32_t main(int32_t argc, char *argv[])
 {
     Connection_information information(argc, argv);
 
@@ -56,7 +55,8 @@ int main(int argc, char *argv[])
     Urg_driver urg;
     if (!urg.open(information.device_or_ip_name(),
                   information.baudrate_or_port_number(),
-                  information.connection_type())) {
+                  information.connection_type()))
+    {
         cout << "Urg_driver::open(): "
              << information.device_or_ip_name() << ": " << urg.what() << endl;
         return 1;
@@ -64,14 +64,19 @@ int main(int argc, char *argv[])
 
     // \~japanese データ取得
     // \~english Gets measurement data
-    enum { Capture_times = 10 };
+    enum
+    {
+        Capture_times = 10
+    };
     urg.start_measurement(Urg_driver::Distance_intensity, Urg_driver::Infinity_times, 0);
-    for (int i = 0; i < Capture_times; ++i) {
-        vector<long> data;
+    for (int32_t i = 0; i < Capture_times; ++i)
+    {
+        vector<int32_t> data;
         vector<unsigned short> intensity;
-        long time_stamp = 0;
+        int32_t time_stamp = 0;
 
-        if (!urg.get_distance_intensity(data, intensity, &time_stamp)) {
+        if (!urg.get_distance_intensity(data, intensity, &time_stamp))
+        {
             cout << "Urg_driver::get_distance(): " << urg.what() << endl;
             return 1;
         }
